// src/app/core/models/studentslog.ts
export interface StudentLog {
    id: number;
    firstName: string;
    lastName: string;
    homegroup: string;
    parentfirstName: string;
    parentlastName: string;
    otherReason: string;
    logDateTime: Date;
    logType: {};
    studentId: number;
    studentLogReasonId: number;
    patronId: number;
  }
  